fixed type issue

+++ ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: idhaimy <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/31 16:59:09 by idhaimy           #+#    #+#             */
-/*   Updated: 2023/10/31 18:37:27 by idhaimy          ###   ########.fr       */
+/*   Updated: 2023/11/04 16:15:00 by idhaimy          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void    *ft_memset(void *s, int c, size_t n)
 {
-        int i;
+        size_t i;
 
         i = 0;
         unsigned char *ss = (unsigned char *)s;
